feat(town): add prefab_shops.js with placeShopPrefabsStrict and wire into town_gen.js
